M4 迁移工具基础版：shanshui-cunji-migrate（CSV/mysqldump 全量导入，development 5.16）；列名即 JSON 字段、docid/id 主键、单线程导入+报告（P24）；147 测试全绿 + 冒烟验证 CSV/SQL 导入正确

+++ src/inverted.rs
@@ -524,10 +524,8 @@ mod tests {
         idx.add("b=2", 2);
         idx.add("a=1", 3);
         let terms = idx.iter_terms().unwrap();
-        let map: std::collections::BTreeMap<String, u64> = terms
-            .into_iter()
-            .map(|(t, b)| (t, b.len() as u64))
-            .collect();
+        let map: std::collections::BTreeMap<String, u64> =
+            terms.into_iter().map(|(t, b)| (t, b.len())).collect();
         assert_eq!(map.get("a=1").copied(), Some(2), "跨段+内存合并");
         assert_eq!(map.get("b=2").copied(), Some(1));
     }
